+++ src/Path.jl
@@ -81,12 +81,16 @@ Base.getindex(cp::Path, r::Union{LinRange, UnitRange, StepRangeLen}) = getindex(
 Base.lastindex(cp::Path) = length(cp.path)
 
 function Base.show(io::IO, cp::Path)
-    println(io, "Path([")
-    for p in cp.path[1:end-1]
-        println(io, " $(p),")
+    if iszero(length(cp))
+        print(io, "Path is empty")
+    else
+        println(io, "Path([")
+        for p in cp.path[1:end-1]
+            println(io, " $(p),")
+        end
+        println(io, " $(cp.path[end])")
+        println(io, "])\n")
     end
-    println(io, " $(cp.path[end])")
-    println(io, "])\n")
 end
 
 drawpath(cpe::PathMove) = move(cpe.pt1)
@@ -225,4 +229,205 @@ function BoundingBox(path::Path)
         end
     end
     return bbox
+end
+
+"""
+      get_bezier_points(bps::BezierPathSegment;
+              steps=10)
+
+The flattening: return a list of all the points on the
+Bezier curve, including start and end, using `steps` to
+determine the accuracy.
+"""
+function get_bezier_points(bps::BezierPathSegment;
+        steps=10)
+    return bezier.(range(0.0, 1.0, length=steps), bps...)
+end
+
+"""
+    get_bezier_length(bps::BezierPathSegment;
+        steps=10)
+
+Return the length of a BezierPathSegment, using `steps` to
+determine the accuracy, by stepping  through the curve and
+finding all the points,  and then measuring between them.
+
+This is obviously just an approximation; the maths to do
+it properly is too difficult for me. :(
+"""
+function get_bezier_length(bps::BezierPathSegment;
+        steps=10)
+    LUT  = get_bezier_points(bps, steps=steps)
+    curvelength = 0.0
+    for i in 1:length(LUT) - 1
+        curvelength += distance(LUT[i], LUT[mod1(i + 1, end)])
+    end
+    return curvelength
+end
+
+"""
+    pathlength(path::Path;
+        steps=10)
+
+Return the length of a Path.
+
+The `steps` parameter is used when approximating the length of any curve (Bezier) sections.
+"""
+function pathlength(path::Path;
+        steps=10)
+    currentpoint = O
+    firstpoint = currentpoint
+    plength = 0.0
+    for pathelement in path
+        if pathelement isa PathMove # pt1
+            firstpoint = pathelement.pt1
+            currentpoint = pathelement.pt1
+        elseif pathelement isa PathCurve # pt1 pt2  pt3
+            plength += get_bezier_length(BezierPathSegment(currentpoint, pathelement.pt1, pathelement.pt2, pathelement.pt3), steps=steps)
+            currentpoint = pathelement.pt3
+        elseif pathelement isa PathLine  # pt1
+            plength += distance(currentpoint, pathelement.pt1)
+            currentpoint = pathelement.pt1
+        elseif pathelement isa PathClose
+            # I think Close is just drawing to the point established by previous Move...
+            plength += distance(firstpoint, currentpoint)
+        end
+    end
+    return plength
+end
+
+"""
+    drawpath(path::Path, k::Real;
+        steps=10, # used when approximating Bezier curve segments
+        action=:none,
+        startnewpath=true,
+        pathlength = 0.0)
+
+Draw the path in `path` starting at the beginning and
+stopping at `k` between 0 and 1. So if `k` is 0.5, half the
+path is drawn.
+
+The function calculates the length of the entire path before
+drawing it. If you want to draw a large path more than once,
+it might be more efficient to calculate the length of the
+path first, and provide it to the `pathlength` keyword.
+
+Returns the last point processed.
+"""
+function drawpath(path::Path, k::Real;
+        steps=10, # used when approximating Bezier curve segments
+        action=:none,
+        startnewpath=true,
+        pathlength = 0.0)
+
+    if iszero(pathlength)
+       pathlength = Luxor.pathlength(path)
+    end
+    requiredlength = k * pathlength
+    currentlength = 0
+    startnewpath && newpath()
+    # firstpoint is the point we will return to for a Close
+    # currentpoint is Cairo's current point
+    # mostrecentpoint is the point we last visited
+    currentpoint = mostrecentpoint = firstpoint = O
+    for pathelement in path
+        if pathelement isa PathMove # pt1
+            currentpoint = firstpoint = pathelement.pt1
+            drawpath(pathelement)
+        elseif pathelement isa PathCurve # pt1 pt2  pt3
+            plength = Luxor.get_bezier_length(BezierPathSegment(currentpoint, pathelement.pt1, pathelement.pt2, pathelement.pt3), steps=steps)
+            currentlength += plength
+            if currentlength > requiredlength
+                # we mustn't draw all of this curve, since it overshoots
+                overshoot  = (currentlength - requiredlength) / pathlength
+                # just draw the overshoot fraction of the curve
+                bps = BezierPathSegment(currentpoint, pathelement.pt1, pathelement.pt2, pathelement.pt3)
+                newbezier = bezier.(range(0.0, overshoot, length=steps), bps...)
+                # using the last three, currentpoint is the first pt
+                drawpath(PathCurve(newbezier[2], newbezier[3], newbezier[4]))
+            else
+                drawpath(pathelement)
+            end
+            mostrecentpoint = currentpoint # remember how far we got
+            currentpoint = pathelement.pt3 # update currentpoint
+        elseif pathelement isa PathLine  # pt1
+            plength = distance(currentpoint, pathelement.pt1)
+            currentlength += plength
+            if currentlength > requiredlength
+                # we mustn't draw all of this line, since it overshoots
+                overshoot  = (currentlength - requiredlength) / pathlength
+                # just draw the overshoot fraction of the line
+            else
+                drawpath(pathelement)
+            end
+            mostrecentpoint = currentpoint
+            currentpoint = pathelement.pt1
+        elseif pathelement isa PathClose
+            # I think Close is just drawing to the point established by previous Move...
+            drawpath(PathLine(firstpoint))
+        end
+        currentlength > requiredlength && break
+    end
+    do_action(action)
+    return mostrecentpoint
+end
+
+drawpath(path::Path, k::Real, act::Symbol;
+        steps=10, # used when approximating Bezier curve segments
+        startnewpath=true,
+        pathlength = 0.0) = drawpath(path, k;
+            steps = steps,
+            action = act,
+            startnewpath = startnewpath,
+            pathlength = pathlength)
+
+"""
+    pathsample(path::Path, spacing)
+
+Return a new Path that resamples the `path` such that each
+line and arc of the original path is divided into sections
+that are `spacing` units long.
+"""
+function pathsample(path::Path, spacing)
+    isapprox(spacing, 0.0) && throw(error("pathsample(): space between vertices must not be zero"))
+    newpath = PathElement[]
+    currentpoint = O
+    firstpoint = currentpoint
+    for pathelement in path
+        if pathelement isa PathMove # pt1
+            push!(newpath, pathelement)
+            firstpoint = pathelement.pt1
+            currentpoint = pathelement.pt1
+        elseif pathelement isa PathCurve # pt1 pt2 pt3
+            bps = BezierPathSegment(currentpoint,
+                pathelement.pt1,
+                pathelement.pt2,
+                pathelement.pt3)
+            plength = Luxor.get_bezier_length(bps)
+            _step = spacing/plength
+            for i in range(0.0, 1.0 - _step, step=_step)
+                bezsegment = trimbezier(bps, i, i + _step)
+                push!(newpath, PathCurve(bezsegment.cp1, bezsegment.cp2, bezsegment.p2))
+            end
+            currentpoint = pathelement.pt3
+        elseif pathelement isa PathLine  # pt1
+            plength = distance(currentpoint, pathelement.pt1)
+            _step = spacing/plength
+            for i in range(0.0, 1.0, step=_step)
+                push!(newpath, PathLine(between(currentpoint, pathelement.pt1, i)))
+            end
+            currentpoint = pathelement.pt1
+        elseif pathelement isa PathClose
+            # I think Close is just drawing to the point established by previous Move...
+            plength = distance(currentpoint, firstpoint)
+            if !iszero(plength)
+                _step = spacing/plength
+                for i in range(0.0, 1.0, step=_step)
+                    push!(newpath, PathLine(between(currentpoint, firstpoint, i)))
+                end
+            end
+            push!(newpath, PathLine(firstpoint))
+        end
+    end
+    return Path(newpath)
 end